move testapi to test subfolder
