+++ operator/test/component/common/common.go
@@ -11,7 +11,7 @@ import (
 	klcv1alpha3 "github.com/keptn/lifecycle-toolkit/operator/apis/lifecycle/v1alpha3"
 	apicommon "github.com/keptn/lifecycle-toolkit/operator/apis/lifecycle/v1alpha3/common"
 	controllercommon "github.com/keptn/lifecycle-toolkit/operator/controllers/common"
-	metricsapi "github.com/keptn/lifecycle-toolkit/operator/testapi/metrics/v1alpha3"
+	metricsapi "github.com/keptn/lifecycle-toolkit/operator/test/api/metrics/v1alpha3"
 	. "github.com/onsi/ginkgo/v2"
 	ginkgotypes "github.com/onsi/ginkgo/v2/types"
 	. "github.com/onsi/gomega"
